violent_proyect/requirements.txt
pandas 
requests
kafka-python
google-cloud-bigquery
python-dotenv